+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/ArLocation.kt
@@ -1,3 +1,0 @@
-package org.neidhardt.arlocation
-
-data class ArLocation(val latitude: Double, val longitude: Double)