+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/GlobalPosition.kt
@@ -1,0 +1,3 @@
+package org.neidhardt.arlocation
+
+data class GlobalPosition(val latitude: Double, val longitude: Double)